Separate unit tests for different functions
