+++ controllers/authController.getOrdersController.test.js
@@ -1,0 +1,79 @@
+import orderModel from "../models/orderModel.js";
+import userModel from "../models/userModel.js";
+
+jest.mock("../models/userModel.js");
+
+const req = {
+    body: {}, 
+    user: { _id: "123" }
+}
+
+const res = {
+    json: jest.fn(),
+    status: jest.fn(() => res),
+    send: jest.fn()
+};
+
+const mockProduct1 = {
+    name: "iphone 16",
+    slug: "iphone 16",
+    description: "Apple's latest smartphone.",
+    price: 2000, 
+    category: { name: "Smartphone" },
+    quantity: 3,
+    photo: {
+        data: null, 
+        contentType: "mockedContent"
+    },
+    shipping: true
+};
+
+const mockProduct2 = {
+    name: "s24",
+    slug: "s24",
+    description: "Samsung's latest smartphone.",
+    price: 1000, 
+    category: { name: "Smartphone" },
+    quantity: 5,
+    photo: {
+        data: null, 
+        contentType: "mockedContent"
+    },
+    shipping: true
+};
+
+const mockProduct3 = {
+    name: "M3 Macbook Air",
+    slug: "M3 Macbook Air",
+    description: "Apple's lightest laptop.",
+    price: 3000, 
+    category: { name: "Laptop" },
+    quantity: 10,
+    photo: {
+        data: null, 
+        contentType: "mockedContent"
+    },
+    shipping: true
+};
+
+const mockProduct4 = {
+    name: "Leobog Hi75",
+    slug: "Leobog Hi75",
+    description: "Thocky keyboard.",
+    price: 40, 
+    category: { name: "Mechanical Keyboard" },
+    quantity: 1,
+    shipping: true
+}
+
+// Tests for getOrdersController
+describe("getOrdersController", () => {
+    beforeAll(() => {
+        console.log()
+    })
+
+    beforeEach(() => {
+        jest.clearAllMocks();
+        req.body = JSON.parse(JSON.stringify(newProfile));
+    });
+});
