+++ controllers/authController.updateProfileController.test.js
@@ -1,7 +1,6 @@
 import { hashPassword } from "../helpers/authHelper.js";
-import orderModel from "../models/orderModel";
-import { updateProfileController } from "./authController";
-import userModel from "../models/userModel";
+import { updateProfileController } from "./authController.js";
+import userModel from "../models/userModel.js";
 
 jest.mock("../models/userModel.js");
 
@@ -233,66 +232,3 @@ describe("updateProfileController", () => {
         expect(res.status).toHaveBeenCalledWith(400);
     });
 });
-
-const mockProduct1 = {
-    name: "iphone 16",
-    slug: "iphone 16",
-    description: "Apple's latest smartphone.",
-    price: 2000, 
-    category: { name: "Smartphone" },
-    quantity: 3,
-    photo: {
-        data: null, 
-        contentType: "mockedContent"
-    },
-    shipping: true
-};
-
-const mockProduct2 = {
-    name: "s24",
-    slug: "s24",
-    description: "Samsung's latest smartphone.",
-    price: 1000, 
-    category: { name: "Smartphone" },
-    quantity: 5,
-    photo: {
-        data: null, 
-        contentType: "mockedContent"
-    },
-    shipping: true
-};
-
-const mockProduct3 = {
-    name: "M3 Macbook Air",
-    slug: "M3 Macbook Air",
-    description: "Apple's lightest laptop.",
-    price: 3000, 
-    category: { name: "Laptop" },
-    quantity: 10,
-    photo: {
-        data: null, 
-        contentType: "mockedContent"
-    },
-    shipping: true
-};
-
-const mockProduct4 = {
-    name: "Leobog Hi75",
-    slug: "Leobog Hi75",
-    description: "Thocky keyboard.",
-    price: 40, 
-    category: { name: "Mechanical Keyboard" },
-    quantity: 1,
-    shipping: true
-}
-
-// Tests for getOrdersController
-// describe("getOrdersController", () => {
-//     beforeAll(() => {
-//         console.log()
-//     })
-
-//     beforeEach(() => {
-//         jest.clearAllMocks();
-//     });
-// });
